Refactor Legal block to be a reusable component
- allows for reuse on the landing page
