+++ app/components/ui/LegalBlockConfig.js
@@ -1,0 +1,49 @@
+const LegalBlockConfig = {
+  TITLE: {
+    BEFORE_BLUE_WORD: 'Discover',
+    BLUE_WORD: 'Legal',
+    AFTER_BLUE_WORD: 'Resources',
+  },
+  FIRST_ROW: [
+    {
+      title: 'Housing',
+      content: 'Eviction defense, housing help, homeownership',
+      query: 'Housing+Law',
+      imgClass: 'legal-block-housing',
+    },
+    {
+      title: 'Immigration',
+      content: 'Asylum, SIJS, DACA, T-Visas, U-Visas, VAWA',
+      query: 'Immigration',
+      imgClass: 'legal-block-immigration',
+    },
+    {
+      title: 'Family & Relationships',
+      content: 'Divorce, custory, guardianship, restraining order, T-Visas, etc.',
+      query: 'Family+Law',
+      imgClass: 'legal-block-family',
+    },
+  ],
+  SECOND_ROW: [
+    {
+      title: 'Criminal & Tickets',
+      content: 'Criminal process questions, quality of life, record expungement',
+      query: 'Criminal',
+      imgClass: 'legal-block-criminal',
+    },
+    {
+      title: 'Work, Credit, and Consumer',
+      content: 'Credit/consumer, worker\'s rights, work and family',
+      query: 'Workers',
+      imgClass: 'legal-block-work',
+    },
+    {
+      title: 'Benefits',
+      content: 'CAAP, CalWORKS, CalFresh, MediCal, SSI',
+      query: 'Government+Benefits',
+      imgClass: 'legal-block-benefits',
+    },
+  ],
+};
+
+export default LegalBlockConfig;
