+++ app/styles/components/_landing-page-card.scss
@@ -22,3 +22,8 @@ $card-width: 305px;
   height: 50%;
   width: 305px;
 }
+
+.landing-page-card__content {
+  padding-bottom: 5%;
+  padding-left: 5%;
+}
