+++ app/styles/components/_landing-page-resource-block.scss
@@ -1,5 +1,40 @@
-.legal-block {
+.landing-page-resource-block {
   background-color: #F7F7F7;
+}
+
+.landing-page-resource-block__resources {
+  display: flex;
+  flex-direction: column;
+  margin: 0 auto;
+  padding-top: 50px;
+  padding-bottom: 50px;
+  position: relative;
+  width: 945px;
+  @include r_max($break-tablet-l) {
+    padding: 50px 0;
+  }
+}
+
+.landing-page-resource-block__resources-title {
+  & h1 {
+    display: inline-block;
+    font-size: 21px;
+    padding: 0px 5px 15px 5px;
+    z-index: 1;
+  }
+
+  & .blue-word {
+    color: $color-brand;
+  }
+}
+
+.landing-page-resource-block__cards {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.landing-page-resource-block__cards-first-row, .landing-page-resource-block__cards-second-row {
+  display: flex;
 }
 
 .legal-block__resources-hammer {
@@ -14,41 +49,6 @@
   @include r_max($break-tablet-l) {
     display: none;
   }
-}
-
-.legal-block__resources {
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-  padding-top: 50px;
-  padding-bottom: 50px;
-  position: relative;
-  width: 945px;
-  @include r_max($break-tablet-l) {
-    padding: 50px 0;
-  }
-}
-
-.legal-block__resources-title {
-  & h1 {
-    display: inline-block;
-    font-size: 21px;
-    padding: 0px 5px 15px 5px;
-    z-index: 1;
-  }
-
-  & .blue-word {
-    color: $color-brand;
-  }
-}
-
-.legal-block__cards {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.legal-block__cards-first-row, .legal-block__cards-second-row {
-  display: flex;
 }
 
 .legal-block-housing {
@@ -73,9 +73,4 @@
 
 .legal-block-benefits {
   background-image: url(../assets/img/legal-block-benefits.png);
-}
-
-.landing-page-card__content {
-  padding-bottom: 5%;
-  padding-left: 5%;
 }
