+++ app/styles/main.scss
@@ -27,6 +27,6 @@
 @import 'components/admin';
 @import 'components/white-label';
 @import 'components/filters';
-@import 'components/legal-block';
+@import 'components/landing-page-resource-block';
 @import 'components/landing-page-card';
 @import 'components/partners';
